+++ packages/spacecat-shared-rum-api-client/src/index.d.ts
@@ -97,6 +97,14 @@ export default class RUMAPIClient {
   createExperimentationURL(params?: RUMAPIOptions): string;
 
   /**
+   * Method to return the url composed of params that the rum-sources API is called with.
+   * @param {RUMAPIOptions} params - An object representing the parameters to be included
+   *  for the rum-sources API call.
+   * @returns A string returning the rum-sources url including query parameters.
+   */
+  createConversionURL(params?: RUMAPIOptions): string;
+
+  /**
    * Asynchronous method to return the 404 sources API call response data.
    * @param {RUMAPIOptions} params - An object representing the parameters to be included
    *  for the 404 sources API call.
@@ -129,4 +137,12 @@ export default class RUMAPIClient {
    *  or an array of all domain urls .
    */
   getDomainList(params?: RUMAPIOptions): Promise<Array<string>>;
+
+  /**
+  * Asynchronous method to return the rum-sources API call response data.
+  * @param {RUMAPIOptions} params - An object representing the parameters to be included
+  *  for the rum-sources data API call.
+  * @returns A Promise resolving to the conversion response data.
+  */
+  getConversionData(params?: RUMAPIOptions): Promise<Array<string>>;
 }