+++ packages/spacecat-shared-ims-client/CHANGELOG.md
@@ -1,3 +1,45 @@
+# [@adobe/spacecat-shared-ims-client-v1.3.4](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ims-client-v1.3.3...@adobe/spacecat-shared-ims-client-v1.3.4) (2024-04-17)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/helix-universal to v4.5.2 ([#211](https://github.com/adobe/spacecat-shared/issues/211)) ([e261e67](https://github.com/adobe/spacecat-shared/commit/e261e677558a6ee165cf0d560625cf665276e144))
+
+# [@adobe/spacecat-shared-ims-client-v1.3.3](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ims-client-v1.3.2...@adobe/spacecat-shared-ims-client-v1.3.3) (2024-04-16)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#202](https://github.com/adobe/spacecat-shared/issues/202)) ([d5d20a8](https://github.com/adobe/spacecat-shared/commit/d5d20a8d606c7589fb794d6c6e3319541a3e6439))
+
+# [@adobe/spacecat-shared-ims-client-v1.3.2](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ims-client-v1.3.1...@adobe/spacecat-shared-ims-client-v1.3.2) (2024-04-12)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/helix-universal to v4.5.1 ([#207](https://github.com/adobe/spacecat-shared/issues/207)) ([985cdba](https://github.com/adobe/spacecat-shared/commit/985cdbaf6c2c354d73f91e94d38244b6eda055ab))
+
+# [@adobe/spacecat-shared-ims-client-v1.3.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ims-client-v1.3.0...@adobe/spacecat-shared-ims-client-v1.3.1) (2024-04-10)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/fetch to v4.1.2 ([#206](https://github.com/adobe/spacecat-shared/issues/206)) ([1f5a638](https://github.com/adobe/spacecat-shared/commit/1f5a638b8ded5a7511ecf8a8f3589cfebe7ce29e))
+
+# [@adobe/spacecat-shared-ims-client-v1.3.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ims-client-v1.2.2...@adobe/spacecat-shared-ims-client-v1.3.0) (2024-03-25)
+
+
+### Features
+
+* Enable retrieval of a user's email address from IMS given an access token [#193](https://github.com/adobe/spacecat-shared/issues/193) ([#194](https://github.com/adobe/spacecat-shared/issues/194)) ([da7735a](https://github.com/adobe/spacecat-shared/commit/da7735a9bf40a3c21d777a49d76ec564a08e07b6))
+
+# [@adobe/spacecat-shared-ims-client-v1.2.2](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ims-client-v1.2.1...@adobe/spacecat-shared-ims-client-v1.2.2) (2024-03-23)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#197](https://github.com/adobe/spacecat-shared/issues/197)) ([776c657](https://github.com/adobe/spacecat-shared/commit/776c657b3ebb2d5b2cd2379aed82d3b9a85ded51))
+
 # [@adobe/spacecat-shared-ims-client-v1.2.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ims-client-v1.2.0...@adobe/spacecat-shared-ims-client-v1.2.1) (2024-03-09)